+++ packages/fluence-js/src/__test__/unit/compiler/v2.spec.ts.obsolete
