+++ packages/fluence-js/src/internal/_aqua/services.ts
@@ -6,9 +6,9 @@
  * Aqua version: 0.7.7-362
  *
  */
-import { FluencePeer } from '../..';
+import { FluencePeer } from '../FluencePeer';
 import type { CallParams$$ } from '../../internal/compilerSupport/v4';
-import { registerService$$ } from '../../internal/compilerSupport/v4';
+import { registerServiceEx } from '../../internal/compilerSupport/v3impl/registerService';
 
 // Services
 
@@ -26,113 +26,114 @@ export interface SigDef {
         callParams: CallParams$$<'signature' | 'data'>,
     ) => boolean | Promise<boolean>;
 }
-export function registerSig(service: SigDef): void;
-export function registerSig(serviceId: string, service: SigDef): void;
-export function registerSig(peer: FluencePeer, service: SigDef): void;
-export function registerSig(peer: FluencePeer, serviceId: string, service: SigDef): void;
 
-export function registerSig(...args: any) {
-    registerService$$(args, {
-        defaultServiceId: 'sig',
-        functions: {
-            tag: 'labeledProduct',
-            fields: {
-                get_peer_id: {
-                    tag: 'arrow',
-                    domain: {
-                        tag: 'nil',
-                    },
-                    codomain: {
-                        tag: 'unlabeledProduct',
-                        items: [
-                            {
-                                tag: 'scalar',
-                                name: 'string',
-                            },
-                        ],
-                    },
-                },
-                sign: {
-                    tag: 'arrow',
-                    domain: {
-                        tag: 'labeledProduct',
-                        fields: {
-                            data: {
-                                tag: 'array',
-                                type: {
+export function registerSig(peer: FluencePeer, serviceId: string, service: any) {
+    registerServiceEx(
+        peer,
+        {
+            defaultServiceId: 'sig',
+            functions: {
+                tag: 'labeledProduct',
+                fields: {
+                    get_peer_id: {
+                        tag: 'arrow',
+                        domain: {
+                            tag: 'nil',
+                        },
+                        codomain: {
+                            tag: 'unlabeledProduct',
+                            items: [
+                                {
                                     tag: 'scalar',
-                                    name: 'u8',
+                                    name: 'string',
+                                },
+                            ],
+                        },
+                    },
+                    sign: {
+                        tag: 'arrow',
+                        domain: {
+                            tag: 'labeledProduct',
+                            fields: {
+                                data: {
+                                    tag: 'array',
+                                    type: {
+                                        tag: 'scalar',
+                                        name: 'u8',
+                                    },
                                 },
                             },
                         },
-                    },
-                    codomain: {
-                        tag: 'unlabeledProduct',
-                        items: [
-                            {
-                                tag: 'struct',
-                                name: 'SignResult',
-                                fields: {
-                                    error: {
-                                        tag: 'option',
-                                        type: {
-                                            tag: 'scalar',
-                                            name: 'string',
-                                        },
-                                    },
-                                    signature: {
-                                        tag: 'option',
-                                        type: {
-                                            tag: 'array',
+                        codomain: {
+                            tag: 'unlabeledProduct',
+                            items: [
+                                {
+                                    tag: 'struct',
+                                    name: 'SignResult',
+                                    fields: {
+                                        error: {
+                                            tag: 'option',
                                             type: {
                                                 tag: 'scalar',
-                                                name: 'u8',
+                                                name: 'string',
                                             },
                                         },
-                                    },
-                                    success: {
-                                        tag: 'scalar',
-                                        name: 'bool',
+                                        signature: {
+                                            tag: 'option',
+                                            type: {
+                                                tag: 'array',
+                                                type: {
+                                                    tag: 'scalar',
+                                                    name: 'u8',
+                                                },
+                                            },
+                                        },
+                                        success: {
+                                            tag: 'scalar',
+                                            name: 'bool',
+                                        },
                                     },
                                 },
-                            },
-                        ],
+                            ],
+                        },
                     },
-                },
-                verify: {
-                    tag: 'arrow',
-                    domain: {
-                        tag: 'labeledProduct',
-                        fields: {
-                            signature: {
-                                tag: 'array',
-                                type: {
-                                    tag: 'scalar',
-                                    name: 'u8',
+                    verify: {
+                        tag: 'arrow',
+                        domain: {
+                            tag: 'labeledProduct',
+                            fields: {
+                                signature: {
+                                    tag: 'array',
+                                    type: {
+                                        tag: 'scalar',
+                                        name: 'u8',
+                                    },
                                 },
-                            },
-                            data: {
-                                tag: 'array',
-                                type: {
-                                    tag: 'scalar',
-                                    name: 'u8',
+                                data: {
+                                    tag: 'array',
+                                    type: {
+                                        tag: 'scalar',
+                                        name: 'u8',
+                                    },
                                 },
                             },
                         },
+                        codomain: {
+                            tag: 'unlabeledProduct',
+                            items: [
+                                {
+                                    tag: 'scalar',
+                                    name: 'bool',
+                                },
+                            ],
+                        },
                     },
-                    codomain: {
-                        tag: 'unlabeledProduct',
-                        items: [
-                            {
-                                tag: 'scalar',
-                                name: 'bool',
-                            },
-                        ],
-                    },in the sens of before 
                 },
             },
         },
-    });
+        serviceId,
+        service,
+    );
 }
 
 // Functions
