+++ packages/fluence-js/src/internal/compilerSupport/v3impl/callFunction.ts
@@ -1,6 +1,5 @@
 import { FnConfig, FunctionCallDef } from './interface';
 import { FluencePeer } from '../../FluencePeer';
-import { Fluence } from '../../../index';
 
 import {
     injectRelayService,
@@ -20,11 +19,11 @@ import {
  * @param def - function definition generated by the Aqua compiler
  * @param script - air script with function execution logic generated by the Aqua compiler
  */
-export function callFunction(rawFnArgs: Array<any>, def: FunctionCallDef, script: string) {
-    const { args, peer, config } = extractArgs(rawFnArgs, def);
+// export function callFunction(rawFnArgs: Array<any>, def: FunctionCallDef, script: string) {
+//     const { args, peer, config } = extractArgs(rawFnArgs, def);
 
-    callFunctionEx(peer, def, script, args, config);
-}
+//     callFunctionEx(peer, def, script, args, config);
+// }
 
 export const callFunctionEx = (
     peer: FluencePeer,
@@ -109,35 +108,35 @@ const isReturnTypeVoid = (def: FunctionCallDef) => {
  * This function select the appropriate configuration and returns
  * arguments in a structured way of: { peer, config, args }
  */
-const extractArgs = (
-    args: any[],
-    def: FunctionCallDef,
-): {
-    peer: FluencePeer;
-    config?: FnConfig;
-    args: any[];
-} => {
-    const numberOfExpectedArgs = getArgumentTypes(def).length;
+// const extractArgs = (
+//     args: any[],
+//     def: FunctionCallDef,
+// ): {
+//     peer: FluencePeer;
+//     config?: FnConfig;
+//     args: any[];
+// } => {
+//     const numberOfExpectedArgs = getArgumentTypes(def).length;
 
-    let peer: FluencePeer;
-    let structuredArgs: any[];
-    let config: any;
-    if (FluencePeer.isInstance(args[0])) {
-        peer = args[0];
-        structuredArgs = args.slice(1, numberOfExpectedArgs + 1);
-        config = args[numberOfExpectedArgs + 1];
-    } else {
-        peer = Fluence.getPeer();
-        structuredArgs = args.slice(0, numberOfExpectedArgs);
-        config = args[numberOfExpectedArgs];
-    }
+//     let peer: FluencePeer;
+//     let structuredArgs: any[];
+//     let config: any;
+//     if (FluencePeer.isInstance(args[0])) {
+//         peer = args[0];
+//         structuredArgs = args.slice(1, numberOfExpectedArgs + 1);
+//         config = args[numberOfExpectedArgs + 1];
+//     } else {
+//         peer = Fluence.getPeer();
+//         structuredArgs = args.slice(0, numberOfExpectedArgs);
+//         config = args[numberOfExpectedArgs];
+//     }
 
-    return {
-        peer: peer,
-        config: config,
-        args: structuredArgs,
-    };
-};
+//     return {
+//         peer: peer,
+//         config: config,
+//         args: structuredArgs,
+//     };
+// };
 
 const getArgumentTypes = (def: FunctionCallDef) => {
     if (def.arrow.domain.tag !== 'labeledProduct') {
