+++ packages/fluence-js/src/internal/compilerSupport/v3impl/registerService.ts
@@ -1,5 +1,4 @@
-import { FluencePeer } from '../../FluencePeer';
-import { Fluence } from '../../../index';
+import type { FluencePeer } from '../../FluencePeer';
 import { ServiceDef } from './interface';
 import { registerGlobalService, userHandlerService } from './services';
 
@@ -10,13 +9,13 @@ import { registerGlobalService, userHandlerService } from './services';
  * @param args - raw arguments passed by user to the generated function
  * @param def - service definition generated by the Aqua compiler
  */
-export function registerService(args: any[], def: ServiceDef) {
-    const { peer, service, serviceId } = extractArgs(args, def.defaultServiceId);
+// export function registerService(args: any[], def: ServiceDef) {
+//     const { peer, service, serviceId } = extractArgs(args, def.defaultServiceId);
 
-    registerServiceEx(peer, service, serviceId);
-}
+//     registerServiceEx(peer, def, service, serviceId);
+// }
 
-export const registerServiceEx = (peer: FluencePeer, serviceId: string, service: any) => {
+export const registerServiceEx = (peer: FluencePeer, def: ServiceDef, serviceId: string, service: any) => {
     if (!peer.getStatus().isInitialized) {
         throw new Error(
             'Could not register the service because the peer is not initialized. Are you passing the wrong peer to the register function?',
@@ -57,43 +56,43 @@ export const registerServiceEx = (peer: FluencePeer, serviceId: string, service:
  * This function select the appropriate configuration and returns
  * arguments in a structured way of: { peer, serviceId, service }
  */
-const extractArgs = (
-    args: any[],
-    defaultServiceId?: string,
-): { peer: FluencePeer; serviceId: string; service: any } => {
-    let peer: FluencePeer;
-    let serviceId: any;
-    let service: any;
-    if (FluencePeer.isInstance(args[0])) {
-        peer = args[0];
-    } else {
-        peer = Fluence.getPeer();
-    }
+// const extractArgs = (
+//     args: any[],
+//     defaultServiceId?: string,
+// ): { peer: FluencePeer; serviceId: string; service: any } => {
+//     let peer: FluencePeer;
+//     let serviceId: any;
+//     let service: any;
+//     if (FluencePeer.isInstance(args[0])) {
+//         peer = args[0];
+//     } else {
+//         peer = Fluence.getPeer();
+//     }
 
-    if (typeof args[0] === 'string') {
-        serviceId = args[0];
-    } else if (typeof args[1] === 'string') {
-        serviceId = args[1];
-    } else {
-        serviceId = defaultServiceId;
-    }
+//     if (typeof args[0] === 'string') {
+//         serviceId = args[0];
+//     } else if (typeof args[1] === 'string') {
+//         serviceId = args[1];
+//     } else {
+//         serviceId = defaultServiceId;
+//     }
 
-    // Figuring out which overload is the service.
-    // If the first argument is not Fluence Peer and it is an object, then it can only be the service def
-    // If the first argument is peer, we are checking further. The second argument might either be
-    // an object, that it must be the service object
-    // or a string, which is the service id. In that case the service is the third argument
-    if (!FluencePeer.isInstance(args[0]) && typeof args[0] === 'object') {
-        service = args[0];
-    } else if (typeof args[1] === 'object') {
-        service = args[1];
-    } else {
-        service = args[2];
-    }
+//     // Figuring out which overload is the service.
+//     // If the first argument is not Fluence Peer and it is an object, then it can only be the service def
+//     // If the first argument is peer, we are checking further. The second argument might either be
+//     // an object, that it must be the service object
+//     // or a string, which is the service id. In that case the service is the third argument
+//     if (!FluencePeer.isInstance(args[0]) && typeof args[0] === 'object') {
+//         service = args[0];
+//     } else if (typeof args[1] === 'object') {
+//         service = args[1];
+//     } else {
+//         service = args[2];
+//     }
 
-    return {
-        peer: peer,
-        serviceId: serviceId,
-        service: service,
-    };
-};
+//     return {
+//         peer: peer,
+//         serviceId: serviceId,
+//         service: service,
+//     };
+// };
